refactor: moved part of logic back to App
