+++ src/components/ItemList/index.tsx
@@ -1,16 +1,18 @@
-import Card, { ItemProps } from '../Card';
+import { BeerDetails } from '../../api/BeerAPI';
+import Card from '../Card';
 
 export interface ItemListProps {
-  items: Array<ItemProps>;
+  items: Array<BeerDetails>;
+  selectItemHandler: (itemId: number) => void;
 }
 
 const ItemList = (props: ItemListProps) => {
-  const { items } = props;
+  const { items, selectItemHandler } = props;
   return (
     <>
       <div style={{ display: 'flex', flexFlow: 'row wrap' }}>
         {items.map((card) => (
-          <Card key={card.id} {...card} />
+          <Card key={card.id} {...card} selectItemHandler={selectItemHandler} />
         ))}
       </div>
     </>
