+++ src/utils/upsertSearchParam.ts
@@ -1,12 +1,16 @@
+import removeSearchParam from './removeSearchParam';
+
 const upsertSearchParam = (
   searchParams: URLSearchParams,
   paramName: string,
   paramValue: string
 ): URLSearchParams => {
+  const updatedSearchParams: URLSearchParams = removeSearchParam(
+    searchParams,
+    paramName
+  );
   const updatedKeyValueParamPairs = [
-    ...Array.from(searchParams.entries()).filter(
-      (keyValuePair: string[]) => keyValuePair[0] !== paramName
-    ),
+    ...Array.from(updatedSearchParams.entries()),
     [paramName, paramValue],
   ];
   return new URLSearchParams(updatedKeyValueParamPairs);
